fix(storage): mark legacy history migration complete
